Use way too much CSS to refactor and redesign manager's order view
