+++ frontend/src/components/AACBoard.js
@@ -32,11 +32,6 @@ function AACBoard({
         <div>
             <SelectedItemsDisplay selectedItems={selectedItems} onDelete={onDeleteItem} onClear={onClearAll} onPlayAll={onPlayAll} />
             <ItemGrid items={menu} onClick={handleClick} />
-            <img
-                src={customerImage ?? "/images/customers/empty.png"}
-                alt="Customer"
-                className="manager-image"
-            />
         </div>
     );
 }
