+++ frontend/src/components/BurgerDisplay.jsx
@@ -1,28 +1,18 @@
 import React from "react";
 import "./BurgerBuilder";
 
-function BurgerDisplay({ imagePaths }) {
+export default function BurgerDisplay({ imagePaths }) {
     return (
         <div className="BurgerDisplay">
-            <img
-                className="GrillImage"
-                src="/images/kitchen.png"
-                alt="grill station"
-            />
-
-            <div className="BurgerStack">
-                {imagePaths.map((path, idx) => (
-                    <img
-                        key={idx}
-                        className="IngredientOnGrill"
-                        src={path}
-                        alt={`ingredient-${idx}`}
-                        style={{ zIndex: idx + 1 }}
-                    />
-                ))}
-            </div>
+            {imagePaths.map((path, idx) => (
+                <img
+                    key={idx}
+                    className="IngredientOnGrill"
+                    src={path}
+                    alt={`ingredient-${idx}`}
+                    style={{ zIndex: idx + 1 }}
+                />
+            ))}
         </div>
     );
 }
-
-export default BurgerDisplay;
