+++ frontend/src/components/BurgerStation.jsx
@@ -1,0 +1,16 @@
+import BurgerDisplay from "./BurgerDisplay";
+
+export default function BurgerStation({ imagePaths }) {
+    return (
+        <div className="BurgerStation">
+            <img
+                className="GrillImage"
+                src="/images/kitchen.png"
+                alt="grill station"
+            />
+            <div className="BurgerStack">
+                <BurgerDisplay imagePaths={imagePaths} />
+            </div>
+        </div>
+    );
+}
